refactor: Move experimental.modelobs to support.modelobs (#2140)
