+++ src/main/java/spoon/support/StandardEnvironment.java
@@ -38,8 +38,8 @@ import spoon.compiler.Environment;
 import spoon.compiler.InvalidClassPathException;
 import spoon.compiler.SpoonFile;
 import spoon.compiler.SpoonFolder;
-import spoon.experimental.modelobs.EmptyModelChangeListener;
-import spoon.experimental.modelobs.FineModelChangeListener;
+import spoon.support.modelobs.EmptyModelChangeListener;
+import spoon.support.modelobs.FineModelChangeListener;
 import spoon.processing.FileGenerator;
 import spoon.processing.ProblemFixer;
 import spoon.processing.ProcessingManager;
