+++ src/main/java/spoon/support/modelobs/ActionBasedChangeListener.java
@@ -14,13 +14,13 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.experimental.modelobs;
+package spoon.support.modelobs;
 
-import spoon.experimental.modelobs.action.Action;
-import spoon.experimental.modelobs.action.AddAction;
-import spoon.experimental.modelobs.action.DeleteAction;
-import spoon.experimental.modelobs.action.DeleteAllAction;
-import spoon.experimental.modelobs.action.UpdateAction;
+import spoon.support.modelobs.action.Action;
+import spoon.support.modelobs.action.AddAction;
+import spoon.support.modelobs.action.DeleteAction;
+import spoon.support.modelobs.action.DeleteAllAction;
+import spoon.support.modelobs.action.UpdateAction;
 
 /**
  * notifies all change on the AST
