+++ src/main/java/spoon/support/modelobs/context/MapContext.java
@@ -14,42 +14,45 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.experimental.modelobs.action;
+package spoon.support.modelobs.context;
 
-import spoon.experimental.modelobs.context.Context;
+import spoon.reflect.declaration.CtElement;
+import spoon.reflect.path.CtRole;
+
+import java.util.Map;
 
 /**
- * defines the update action
- * @param <T>
+ * defines the map context
+ * @param <K>
+ * @param <V>
  */
-public class UpdateAction<T> extends Action {
-	private final T oldValue;
-	private final T newValue;
+public class MapContext<K, V> extends Context {
+	private final Map<K, V> map;
+	private  K key;
 
-	public UpdateAction(Context context, T newValue, T oldValue) {
-		super(context);
-		this.oldValue = oldValue;
-		this.newValue = newValue;
+	public MapContext(CtElement element, CtRole role, Map<K, V> map) {
+		super(element, role);
+		this.map = map;
 	}
 
-	@Override
-	public T getChangedValue() {
-		return getNewValue();
-	}
-
-	/**
-	 * the new value in the model
-	 * @return the new value
-	 */
-	public T getNewValue() {
-		return newValue;
+	public MapContext(CtElement element, CtRole role, Map<K, V> map, K key) {
+		this(element, role, map);
+		this.key = key;
 	}
 
 	/**
-	 * the old value in the model
-	 * @return the old value
+	 * get the changed key
+	 * @return the changed key
 	 */
-	public T getOldValue() {
-		return oldValue;
+	public K getKey() {
+		return key;
+	}
+
+	/**
+	 * the changed map
+	 * @return the changed map
+	 */
+	public Map<K, V> getMap() {
+		return map;
 	}
 }
