+++ src/main/java/spoon/support/modelobs/context/ObjectContext.java
@@ -14,34 +14,14 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.experimental.modelobs.context;
+package spoon.support.modelobs.context;
 
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.path.CtRole;
 
-import java.util.List;
+public class ObjectContext extends Context {
 
-/**
- * defines a list context
- */
-public class ListContext extends CollectionContext<List<?>> {
-	private final int position;
-
-	public ListContext(CtElement element, CtRole role, List<?> original) {
-		this(element, role, original, -1);
-	}
-
-	public ListContext(CtElement element, CtRole role, List<?> original, int position) {
-		super(element, role, original);
-		this.position = position;
-	}
-
-	/**
-	 * the position where the change has been made (returns -1 if no position is defined).
-	 *
-	 * @return the position of the change
-	 */
-	public int getPosition() {
-		return position;
+	public ObjectContext(CtElement ctElement, CtRole role) {
+		super(ctElement, role);
 	}
 }
