+++ src/test/java/spoon/test/parent/ContractOnSettersParametrizedTest.java
@@ -4,8 +4,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import spoon.SpoonException;
-import spoon.experimental.modelobs.ActionBasedChangeListenerImpl;
-import spoon.experimental.modelobs.action.Action;
+import spoon.support.modelobs.ActionBasedChangeListenerImpl;
+import spoon.support.modelobs.action.Action;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.declaration.CtPackage;
